1.swf动态表名适配
