--- src/main/java/com/om/DataMagic/infrastructure/pgDB/service/WatchService.java
@@ -1,25 +1,0 @@
-/*
- This project is licensed under the Mulan PSL v2.
- You can use this software according to the terms and conditions of the Mulan PSL v2.
- You may obtain a copy of Mulan PSL v2 at:
-     http://license.coscl.org.cn/MulanPSL2
- THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND,
- EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT,
- MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
- See the Mulan PSL v2 for more details.
- Created: 2025
-*/
-
-package com.om.DataMagic.infrastructure.pgDB.service;
-
-import com.baomidou.mybatisplus.extension.service.IService;
-import com.om.DataMagic.infrastructure.pgDB.dataobject.WatchDO;
-
-/**
- * watch service 接口.
- *
- * @author zhaoyan
- * @since 2025-01-14
- */
-public interface WatchService extends IService<WatchDO> {
-}
