+++ src/main/java/com/om/DataMagic/infrastructure/pgDB/service/platform/StarService.java
@@ -10,7 +10,7 @@
  Created: 2025
 */
 
-package com.om.DataMagic.infrastructure.pgDB.service;
+package com.om.DataMagic.infrastructure.pgDB.service.platform;
 
 import com.baomidou.mybatisplus.extension.service.IService;
 import com.om.DataMagic.infrastructure.pgDB.dataobject.StarDO;
