+++ src/main/java/com/om/DataMagic/infrastructure/pgDB/service/platform/WatchServiceImpl.java
@@ -10,12 +10,15 @@
  Created: 2025
 */
 
-package com.om.DataMagic.infrastructure.pgDB.service;
+package com.om.DataMagic.infrastructure.pgDB.service.platform;
 
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.om.DataMagic.infrastructure.pgDB.dataobject.WatchDO;
 import com.om.DataMagic.infrastructure.pgDB.mapper.WatchDOMapper;
 import org.springframework.stereotype.Service;
+
+import java.util.Collection;
+import java.util.List;
 
 /**
  * watch service 实现类.
@@ -25,4 +28,21 @@ import org.springframework.stereotype.Service;
  */
 @Service
 public class WatchServiceImpl extends ServiceImpl<WatchDOMapper, WatchDO> implements WatchService {
+    /**
+     * @param entityList sava or update .
+     * @return
+     */
+    @Override
+    public boolean saveOrUpdateBatch(Collection<WatchDO> entityList) {
+        return super.saveOrUpdateBatch(entityList);
+    }
+
+    /**
+     * query return list.
+     * @return list
+     */
+    @Override
+    public List<WatchDO> list() {
+        return super.list();
+    }
 }
