+++ src/main/java/com/om/DataMagic/process/codePlatform/gitcode/GitCodeStarProcess.java
@@ -19,7 +19,7 @@ import com.om.DataMagic.infrastructure.pgDB.converter.StarConverter;
 import com.om.DataMagic.infrastructure.pgDB.dataobject.RepoDO;
 import com.om.DataMagic.infrastructure.pgDB.dataobject.StarDO;
 import com.om.DataMagic.infrastructure.pgDB.service.platform.RepoService;
-import com.om.DataMagic.infrastructure.pgDB.service.StarService;
+import com.om.DataMagic.infrastructure.pgDB.service.platform.StarService;
 import com.om.DataMagic.process.DriverManager;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
