+++ src/main/java/com/om/DataMagic/process/codePlatform/gitee/GiteeForkProcess.java
@@ -18,7 +18,7 @@ import com.om.DataMagic.common.constant.CodePlatformEnum;
 import com.om.DataMagic.infrastructure.pgDB.converter.ForkConverter;
 import com.om.DataMagic.infrastructure.pgDB.dataobject.ForkDO;
 import com.om.DataMagic.infrastructure.pgDB.dataobject.RepoDO;
-import com.om.DataMagic.infrastructure.pgDB.service.ForkService;
+import com.om.DataMagic.infrastructure.pgDB.service.platform.ForkService;
 import com.om.DataMagic.infrastructure.pgDB.service.platform.RepoService;
 import com.om.DataMagic.process.DriverManager;
 import org.springframework.beans.factory.annotation.Autowired;
